Moves out social media tests to new class, and created a seperate test class for them.
